Exclude abstract class from being loaded
