+++ src/nodes/draw/abstract.tsx
@@ -1,6 +1,7 @@
 import PPNode from '../../classes/NodeClass';
 import PPGraph from '../../classes/GraphClass';
 import Socket from '../../classes/SocketClass';
+import HybridNode from '../../classes/HybridNode';
 import {
   NODE_TYPE_COLOR,
   PIXI_PIVOT_OPTIONS,
